# gh-119109: functools.partial.fallforward fix