https://github.com/python/cpython/issues/119109

I went with option 1. It is cleaner, faster and has a natural flow.

Before:
```python
from functools import partial

def func(a, b, c=0):
    return a - b - c

p1 = partial(func, 1)

print(p1(2))    # -1
%timeit p1(2)   # 100 ns

p1.keywords['c'] = 1
print(p1(2))    # -2
%timeit p1(2)   # 234 ns

del p1.keywords['c']
print(p1(2))    # -1
%timeit p1(2)   # 144 ns
```

After:
```python
p1 = partial(func, 1)

print(p1(2))    # -1
%timeit p1(2)   # 100 ns

p1.keywords['c'] = 1
print(p1(2))    # -2
%timeit p1(2)   # 234 ns

del p1.keywords['c']
print(p1(2))    # -1
%timeit p1(2)   # 100 ns
```

<!-- gh-issue-number: gh-119109 -->
* Issue: gh-119109
<!-- /gh-issue-number -->